Oodo_R-main/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
numpy>=1.24.0
scipy>=1.11.0
mysql-connector-python>=8.0.0
